+++ crates/idxdb-store/src/chain_data/mod.rs
@@ -14,31 +14,22 @@ use crate::promise::{await_js, await_js_value, await_ok};
 
 mod js_bindings;
 use js_bindings::{
-    idxdb_get_block_headers,
-    idxdb_get_current_blockchain_peaks,
-    idxdb_get_partial_blockchain_nodes,
-    idxdb_get_partial_blockchain_nodes_all,
-    idxdb_get_partial_blockchain_nodes_up_to_inorder_index,
-    idxdb_get_tracked_block_header_numbers,
-    idxdb_get_tracked_block_headers,
-    idxdb_insert_block_header,
-    idxdb_insert_partial_blockchain_nodes,
-    idxdb_prune_irrelevant_blocks,
+    idxdb_get_block_headers, idxdb_get_current_blockchain_peaks,
+    idxdb_get_partial_blockchain_nodes, idxdb_get_partial_blockchain_nodes_all,
+    idxdb_get_partial_blockchain_nodes_up_to_inorder_index, idxdb_get_tracked_block_header_numbers,
+    idxdb_get_tracked_block_headers, idxdb_insert_block_header,
+    idxdb_insert_partial_blockchain_nodes, idxdb_prune_irrelevant_blocks,
 };
 
 mod models;
 use models::{
-    BlockHeaderIdxdbObject,
-    PartialBlockchainNodeIdxdbObject,
-    PartialBlockchainPeaksIdxdbObject,
+    BlockHeaderIdxdbObject, PartialBlockchainNodeIdxdbObject, PartialBlockchainPeaksIdxdbObject,
 };
 
 pub mod utils;
 use utils::{
-    SerializedBlockHeaderData,
-    SerializedPartialBlockchainNodeData,
-    process_partial_blockchain_nodes_from_js_value,
-    serialize_block_header,
+    SerializedBlockHeaderData, SerializedPartialBlockchainNodeData,
+    process_partial_blockchain_nodes_from_js_value, serialize_block_header,
     serialize_partial_blockchain_node,
 };