+++ crates/idxdb-store/src/transaction/utils.rs
@@ -5,10 +5,7 @@ use miden_client::Word;
 use miden_client::note::{BlockNumber, ToInputNoteCommitments};
 use miden_client::store::StoreError;
 use miden_client::transaction::{
-    ExecutedTransaction,
-    TransactionDetails,
-    TransactionRecord,
-    TransactionScript,
+    ExecutedTransaction, TransactionDetails, TransactionRecord, TransactionScript,
     TransactionStatus,
 };
 use miden_client::utils::Serializable;